% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.pvarife_result}
\alias{coef.pvarife_result}
\title{Extract coefficients from a pvarife_result}
\usage{
\method{coef}{pvarife_result}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"pvarife_result"}.}

\item{...}{Ignored.}
}
\value{
Named numeric vector of coefficients.
}
\description{
Extract coefficients from a pvarife_result
}
