% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pvarife_bands}
\alias{plot.pvarife_bands}
\title{Plot impulse response bands}
\usage{
\method{plot}{pvarife_bands}(x, var_names = NULL, normalise_by_h1 = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"pvarife_bands"}.}

\item{var_names}{Character vector of variable names (optional).}

\item{normalise_by_h1}{Logical. If \code{TRUE}, divide all responses by the
first-horizon response of the shock variable (replicates Figure 1 of
Tugan 2021). Default \code{FALSE}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Plot impulse response bands
}
