% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pvarife_irf}
\alias{plot.pvarife_irf}
\title{Plot impulse response functions}
\usage{
\method{plot}{pvarife_irf}(x, var_names = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"pvarife_irf"}.}

\item{var_names}{Character vector of variable names (optional).}

\item{...}{Ignored.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Plot impulse response functions
}
