% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pvarife_result}
\alias{print.pvarife_result}
\title{Print method for pvarife_result}
\usage{
\method{print}{pvarife_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"pvarife_result"}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for pvarife_result
}
