% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.pvarife_result}
\alias{summary.pvarife_result}
\title{Summary method for pvarife_result}
\usage{
\method{summary}{pvarife_result}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"pvarife_result"}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns a list with \code{fit}, \code{avar}.
}
\description{
Summary method for pvarife_result
}
